backend/microservicios_mesas/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
flask_jwt_extended==4.6.0